requirements.txt
SQLAlchemy
python-jose
passlib
python-multipart
langchain
fastapi
uvicorn
python-dotenv
langchain-community
openai